pyproject.toml
[tool.poetry]
name = "django-rest-framework-project"
version = "0.1.0"
description = ""
authors = ["Chera <chera1904@mail.ru>"]
readme = "README.md"
#packages = [{include = "django_rest_framework_project"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2.1"
jinja2 = "^3.1.2"
djangorestframework = "^3.14.0"
djoser = "^2.2.0"
djangorestframework-simplejwt = "^5.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
